backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
numpy>=1.26.0,<3.0.0
Pillow>=10.0.0,<12.0.0
httpx>=0.27.0,<1.0.0
python-dotenv>=1.0.0
